refactor(client): post e2e-test-c rebase on
